+++ src/Core/Dfe.Complete.Domain/Entities/ConversionTasksData.cs
@@ -76,13 +76,13 @@ public class ConversionTasksData : BaseAggregateRoot, IEntity<TaskDataId>
 
     public bool? ChurchSupplementalAgreementCleared { get; set; }
 
-    public bool? ChurchSupplementalAgreementSigned { get; set; }
+    public bool? ChurchSupplementalAgreementSignedTrust { get; set; }
 
     public bool? ChurchSupplementalAgreementSignedDiocese { get; set; }
 
-    public bool? ChurchSupplementalAgreementSaved { get; set; }
+    public bool? ChurchSupplementalAgreementDraftSaved { get; set; }
 
-    public bool? ChurchSupplementalAgreementSent { get; set; }
+    public bool? ChurchSupplementalAgreementFinalSaved { get; set; }
 
     public bool? ChurchSupplementalAgreementSignedSecretaryState { get; set; }