+++ CHANGELOG.md
@@ -23,6 +23,11 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - [277839](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/277839) - Master funding agreement content changes
 ---
 
+### Changed
+- [277820](https://dev.azure.com/dfe-gov-uk/Academies-and-Free-Schools-SIP/_workitems/edit/277820) - Church Supplemental Agreement
+
+---
+
 ## [1.46.1](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2026-05-21.1562) - 2026-05-21
 
 ### Changed